+++ src/RezeptionPatientBearbeiten.java
@@ -121,7 +121,7 @@ public class RezeptionPatientBearbeiten extends JFrame {
                     return;
                 }
                 try {
-                    patient.setSozialversicherungsnummer(Integer.parseInt(svnTextField.getText()));
+                    patient.setSozialversicherungsnummer(Long.parseLong(svnTextField.getText()));
                 } catch (NumberFormatException ex) {
                     SwingUtilities.invokeLater(() ->
                             JOptionPane.showMessageDialog(RezeptionPatientBearbeiten.this,
